45 modificar clase compra para compraitem (#46)
* Clase Comprar modificada para relacionarla con la clase comprarItem

+++ src/AppForSEII2526.API/Models/ComprarItem.cs
@@ -7,7 +7,7 @@
 
         // Permite crear un nuevo ítem de compra asociando un coche y una compra concretos.
         // guarda el precio actual del coche y la cantidad comprada.
-        public ComprarItem(Coche coche, Comprar comprar, double cantidad)
+        public ComprarItem(Coche coche, Comprar comprar, decimal cantidad)
         {
             Coche = coche;
             CocheId = coche.Id;
@@ -27,7 +27,7 @@
         public decimal Precio { get; set; }
 
         [Range(1, int.MaxValue, ErrorMessage = "La cantidad debe ser mayor que 1")]
-        public double Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
     }
 
 }
